fix memory leak
Turns out allocating new three js objects every time we change the
keyboard is asking for bad times and memory leaks.

As a fix, only allocating objects once. When the keyboard is changed,
then set new colors assuming key ordering matches between the three js
scene and the keyboard object. This will lead to problems when we want
to support keyboards in different formats -- right now everything is
hardcoded to be tenkeyless.